Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 int main(){
-    int a,b;
+    float a,b;
     char ch;
-    scanf("%d%d",&a,&b);
+    scanf("%f%f",&a,&b);
     scanf("%c",&ch);
     switch(ch){
         case '+':
-            printf("%d",a+b);
+            printf("%f",a+b);
             break;
         case '-':
-            printf("%d",a-b);
+            printf("%f",a-b);
             break;
         case '*':
-            printf("%d",a*b);
+            printf("%f",a*b);
             break;
         case '/':
             if(b!=0){
-                printf("%f",(float)a/b);
+                printf("%.2f",a/b);
             }
-            else{
+                else{
                 printf("error");
             }
             break;
